pyproject.toml
[project]
name = "hm-dl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gdown>=5.2.0",
    "ipykernel>=6.29.5",
    "kaggle>=1.7.4.5",
    "keras==3.5.0",
    "matplotlib==3.8.0",
    "numpy==1.26.4",
    "pillow>=11.3.0",
    "scikit-learn==1.6.0",
    "seaborn>=0.13.2",
    "tensorflow==2.17.1",
    "tensorflow-hub>=0.16.1",
    "tensorflow-metal>=1.1.0",
    "tf-keras>=2.17.0",
    "torch==2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "transformers==4.47.1",
]
